moves login-form component to folder
