+++ src/app/login/components/login-form/login-form.component.ts
@@ -1,5 +1,5 @@
 import { Component, Output, EventEmitter } from '@angular/core';
-import { User } from '../models/user.interface';
+import { User } from '../../models/user.interface';
 
 @Component({
     selector: 'app-login-form',
