+++ src/app/login/login.module.ts
@@ -6,7 +6,7 @@ import { FormsModule } from '@angular/forms';
 // Containers
 import { LoginComponent } from './containers/login.component';
 // Components
-import { LoginFormComponent } from './components/login-form.component';
+import { LoginFormComponent } from './components/login-form/login-form.component';
 // Services
 import { LoginService } from './services/login.service';
